requirements.txt
Django>=4.0
Pillow>=9.0
requests>=2.25